Fix validateInput
Fixed validateInput to provide an alert if something besides one
of the three allowed choices is made. It then loops back to the
getPlayerInput function so the player can try again. The do...while
loop was chosen to force at least one prompt for input. The ending
on escape or cancel was kept.