Add UI buttons to replace the replay prompt
Removed the annoying pop up prompt asking for a new game. Replaced it
with buttons that appear once the game is over. A new event listener
had to be added for the new buttons. The buttons are hard coded into
the HTML but are hidden unil the game is over and a new class is
added to them.