Center stats with flexbox
Using Flexbox results in a more consistent centering of the text
on the screen. This was accmplished by putting the stats into the
same `div` that held the selection buttons. Initially, this resulted
in the stats being to the side rather than below the buttons. I fixed
that issue by setting flex-wrap to wrap and putting in a `div` with a width of 100vw to force a
line break in the flexbox.

Also minor formatting fixes of button placement on small screens.